Fixed recording of samples before pulse

+++ MeasureGraph/AutoMeasureGraphMin/pulse.h
@@ -3,5 +3,5 @@
 typedef struct pulseData_ {
     float voltage;
     float current;
-    int micros_from_start;
+    unsigned long micros_from_start;
 } pulseData;
